+++ HISTORY.md
@@ -3,8 +3,13 @@ hotpotato
 
 0.7.1
 -----
-Added support for HTTP Sessions (cookie storage, authentication, automatic redirection, etc) - package org.factor45.hotpotato.session
+* Fixed bug in DefaultHttpConnection where HttpHeaders.isKeepAlive was being used on request, rather than response.
+* Added support for HTTP Sessions: way simpler API, perfect for client-side usage (package org.factor45.hotpotato.session)
+  - cookie storage;
+  - authentication;
+  - automatic redirection;
+  - proxy support.
 
 0.7.0
 -----
-First release!
+* First release!